Capstone Project - IBM Data Analysis certificate
To be competitive in the global IT sector, It's essential to keep up with the ever-changing technologies. The project uses data analytics to highlight current and projected trends in the need for skills related to programming languages, databases and other technologies. It also studies the demographics of professionals in the technology sector.
Used Python, SQL, Cognos Dashboard, Excel, Power Point to successfully conduct the analysis.